Update vbrev8.adoc
Changed the description of input and output to clarify that they are still viewed as elements of size SEW. While the operation is performed at the byte level, the elements are still the unit of execution for vl and masking purposes. This is the same as other bit- and byte-wise instructions.